refactor to create socket server class and moved all logging functions into a Utils header
